requirements.txt
litellm
pandas
numpy
streamlit
ragas
langchain
wandb
python-dotenv
pyyaml
openpyxl
reportlab
altair